update de date sizes and add the lower bounds in the map

+++ contracts/Ballot.sol
@@ -1,22 +1,22 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.11;
 
-// TODO: change the add voters way
+// TODO: change the adding voters way
 import "hardhat/console.sol";
 
 contract Ballot {
     struct Voter {
-        uint256 weight; // weight is accumulated by delegation
+        uint32 weight; // weight is accumulated by delegation
         bool voted; // if true, that person already voted
         address delegate; // person delegated to
-        uint256 vote; // index of the voted proposal
+        uint16 vote; // index of the voted proposal
     }
 
     struct Coordinates {
-        uint256 x1;
-        uint256 x2;
-        uint256 y1;
-        uint256 y2;
+        uint32 x1;
+        uint32 x2;
+        uint32 y1;
+        uint32 y2;
     }
 
     // This is a type for a single proposal.
@@ -25,7 +25,7 @@ contract Ballot {
         Coordinates coo;
         bool isGetLands;
         address to;
-        uint256 voteCount; // number of accumulated votes
+        uint32 voteCount; // number of accumulated votes
     }
 
     address public chairperson;
@@ -40,10 +40,10 @@ contract Ballot {
     /// Create a new ballot to choose one of `proposals`.
     constructor(
         address[] memory _voters,
-        uint256 x1_,
-        uint256 x2_,
-        uint256 y1_,
-        uint256 y2_,
+        uint32 x1_,
+        uint32 x2_,
+        uint32 y1_,
+        uint32 y2_,
         bool isGetLands_,
         address to_
     ) {
@@ -51,7 +51,7 @@ contract Ballot {
         chairperson = _voters[0];
 
         //set the voters address
-        for (uint256 i = 0; i < _voters.length; i++) {
+        for (uint32 i = 0; i < _voters.length; i++) {
             voters[_voters[i]].weight = 1;
         }
 
@@ -94,7 +94,7 @@ contract Ballot {
 
     /// Give your vote (including votes delegated to you)
     /// to proposal `proposals[proposal].name`.
-    function vote(uint256 proposal) external {
+    function vote(uint16 proposal) external {
         Voter storage sender = voters[msg.sender];
         require(sender.weight != 0, "Has no right to vote");
         require(!sender.voted, "Already voted.");
@@ -109,9 +109,9 @@ contract Ballot {
 
     /// @dev Computes the winning proposal taking all
     /// previous votes into account.
-    function winningProposal() public view returns (uint256 winningProposal_) {
-        uint256 winningVoteCount = 0;
-        for (uint256 p = 0; p < proposals.length; p++) {
+    function winningProposal() public view returns (uint32 winningProposal_) {
+        uint32 winningVoteCount = 0;
+        for (uint32 p = 0; p < proposals.length; p++) {
             if (proposals[p].voteCount > winningVoteCount) {
                 winningVoteCount = proposals[p].voteCount;
                 winningProposal_ = p;
@@ -134,7 +134,7 @@ contract Ballot {
         isGet = proposals[winningProposal()].isGetLands;
     }
 
-    function getProposal(uint256 proposal) external view returns (Proposal memory prop) {
+    function getProposal(uint32 proposal) external view returns (Proposal memory prop) {
         prop = proposals[proposal];
     }
 }
